Moved main module pages out of subdirectories
