+++ src/modules/main/pages/home.vue
